Rename repo and move subpackages to pkg/
