+++ pkg/bolt/client.go
