+++ pkg/bolt/client_test.go
