+++ pkg/bolt/interface.go
