+++ pkg/message/generic.go
