+++ pkg/websock/README.md
