+++ pkg/websock/utils.go
